+++ warp/context.py
@@ -34,7 +34,7 @@ import warp.config
 
 
 def create_value_func(type):
-    def value_func(args, kwds, templates):
+    def value_func(arg_types, arg_values):
         return type
 
     return value_func
@@ -63,6 +63,8 @@ class Function:
         input_types=None,
         value_type=None,
         value_func=None,
+        export_func=None,
+        dispatch_func=None,
         module=None,
         variadic=False,
         initializer_list_func=None,
@@ -96,13 +98,15 @@ class Function:
         self.namespace = namespace
         self.value_type = value_type
         self.value_func = value_func  # a function that takes a list of args and a list of templates and returns the value type, e.g.: load(array, index) returns the type of value being loaded
+        self.export_func = export_func
+        self.dispatch_func = dispatch_func
         self.input_types = {}
         self.export = export
         self.doc = doc
         self.group = group
         self.module = module
         self.variadic = variadic  # function can take arbitrary number of inputs, e.g.: printf()
-        self.defaults = defaults
+        self.defaults = {} if defaults is None else defaults
         # Function instance for a custom implementation of the replay pass
         self.custom_replay_func = custom_replay_func
         self.native_snippet = native_snippet
@@ -178,6 +182,33 @@ class Function:
         if not skip_adding_overload:
             self.add_overload(self)
 
+        # Store a description of the function's signature that can be used
+        # to resolve a bunch of positional/keyword/variadic arguments against,
+        # in a way that is compatible with Python's semantics.
+        signature_params = []
+        signature_default_param_kind = inspect.Parameter.POSITIONAL_OR_KEYWORD
+        for param_name in self.input_types.keys():
+            if param_name.startswith("**"):
+                param_name = param_name[2:]
+                param_kind = inspect.Parameter.VAR_KEYWORD
+            elif param_name.startswith("*"):
+                param_name = param_name[1:]
+                param_kind = inspect.Parameter.VAR_POSITIONAL
+
+                # Once a variadic argument like `*args` is found, any following
+                # arguments need to be passed using keywords.
+                signature_default_param_kind = inspect.Parameter.KEYWORD_ONLY
+            else:
+                param_kind = signature_default_param_kind
+
+            param = param = inspect.Parameter(
+                param_name,
+                param_kind,
+                default=self.defaults.get(param_name, inspect.Parameter.empty),
+            )
+            signature_params.append(param)
+        self.signature = inspect.Signature(signature_params)
+
         # add to current module
         if module:
             module.register_function(self, skip_adding_overload)
@@ -245,7 +276,7 @@ class Function:
 
         # only export simple types that don't use arrays
         for v in self.input_types.values():
-            if isinstance(v, warp.array) or v in complex_type_hints:
+            if warp.types.is_array(v) or v in complex_type_hints:
                 return False
 
         if type(self.value_type) in sequence_types:
@@ -259,8 +290,14 @@ class Function:
 
         name = "builtin_" + self.key
 
+        # Runtime arguments that are to be passed to the function, not its template signature.
+        if self.export_func is not None:
+            func_args = self.export_func(self.input_types)
+        else:
+            func_args = self.input_types
+
         types = []
-        for t in self.input_types.values():
+        for t in func_args.values():
             types.append(t.__name__)
 
         return "_".join([name, *types])
@@ -297,7 +334,7 @@ class Function:
                     )
                 self.user_overloads[sig] = f
 
-    def get_overload(self, arg_types):
+    def get_overload(self, arg_types, kwarg_types):
         assert not self.is_builtin()
 
         sig = warp.types.get_signature(arg_types, func_name=self.key)
@@ -350,10 +387,16 @@ def call_builtin(func: Function, *params) -> Tuple[bool, Any]:
     # Retrieve the built-in function from Warp's dll.
     c_func = getattr(warp.context.runtime.core, func.mangled_name)
 
+    # Runtime arguments that are to be passed to the function, not its template signature.
+    if func.export_func is not None:
+        func_args = func.export_func(func.input_types)
+    else:
+        func_args = func.input_types
+
     # Try gathering the parameters that the function expects and pack them
     # into their corresponding C types.
     c_params = []
-    for i, (_, arg_type) in enumerate(func.input_types.items()):
+    for i, (_, arg_type) in enumerate(func_args.items()):
         param = params[i]
 
         try:
@@ -483,7 +526,8 @@ def call_builtin(func: Function, *params) -> Tuple[bool, Any]:
                 c_params.append(arg_type._type_(param))
 
     # returns the corresponding ctype for a scalar or vector warp type
-    value_type = func.value_func(None, None, None)
+    value_type = func.value_func(None, None)
+
     if value_type == float:
         value_ctype = ctypes.c_float
     elif value_type == int:
@@ -804,7 +848,7 @@ def func_replay(forward_fn):
                 f"Cannot define custom replay definition for {forward_fn.key} since the provided replay function has generic input arguments."
             )
 
-        f = forward_fn.get_overload(arg_types)
+        f = forward_fn.get_overload(arg_types, {})
         if f is None:
             inputs_str = ", ".join([f"{k}: {v.__name__}" for k, v in args.items()])
             raise RuntimeError(
@@ -816,6 +860,8 @@ def func_replay(forward_fn):
             namespace=f.namespace,
             input_types=f.input_types,
             value_func=f.value_func,
+            export_func=f.export_func,
+            dispatch_func=f.dispatch_func,
             module=f.module,
             skip_reverse_codegen=True,
             skip_adding_overload=True,
@@ -961,6 +1007,8 @@ def add_builtin(
     constraint=None,
     value_type=None,
     value_func=None,
+    export_func=None,
+    dispatch_func=None,
     doc="",
     namespace="wp::",
     variadic=False,
@@ -980,12 +1028,12 @@ def add_builtin(
     # wrap simple single-type functions with a value_func()
     if value_func is None:
 
-        def value_func(args, kwds, templates):
+        def value_func(arg_types, arg_values):
             return value_type
 
     if initializer_list_func is None:
 
-        def initializer_list_func(args, templates):
+        def initializer_list_func(args, return_type):
             return False
 
     if defaults is None:
@@ -993,8 +1041,13 @@ def add_builtin(
 
     # Add specialized versions of this builtin if it's generic by matching arguments against
     # hard coded types. We do this so you can use hard coded warp types outside kernels:
+    if export_func is not None:
+        func_arg_types = export_func(input_types)
+    else:
+        func_arg_types = input_types
+
     generic = False
-    for x in input_types.values():
+    for x in func_arg_types.values():
         if warp.types.type_is_generic(x):
             generic = True
             break
@@ -1002,7 +1055,7 @@ def add_builtin(
     if generic and export:
         # collect the parent type names of all the generic arguments:
         genericset = set()
-        for t in input_types.values():
+        for t in func_arg_types.values():
             if hasattr(t, "_wp_generic_type_hint_"):
                 genericset.add(t._wp_generic_type_hint_)
             elif warp.types.type_is_generic_scalar(t):
@@ -1054,15 +1107,17 @@ def add_builtin(
 
                 typelists.append(l)
 
-            for argtypes in itertools.product(*typelists):
+            for arg_types in itertools.product(*typelists):
+                arg_types = dict(zip(input_types.keys(), arg_types))
+
                 # Some of these argument lists won't work, eg if the function is mul(), we won't be
                 # able to do a matrix vector multiplication for a mat22 and a vec3. The `constraint`
                 # function determines which combinations are valid:
                 if constraint:
-                    if constraint(argtypes) is False:
+                    if constraint(arg_types) is False:
                         continue
 
-                return_type = value_func(argtypes, {}, [])
+                return_type = value_func(arg_types, None)
 
                 # The return_type might just be vector_t(length=3,dtype=wp.float32), so we've got to match that
                 # in the list of hard coded types so it knows it's returning one of them:
@@ -1080,8 +1135,10 @@ def add_builtin(
                 # finally we can generate a function call for these concrete types:
                 add_builtin(
                     key,
-                    input_types=dict(zip(input_types.keys(), argtypes)),
+                    input_types=arg_types,
                     value_type=return_type,
+                    export_func=export_func,
+                    dispatch_func=dispatch_func,
                     doc=doc,
                     namespace=namespace,
                     variadic=variadic,
@@ -1091,6 +1148,7 @@ def add_builtin(
                     hidden=True,
                     skip_replay=skip_replay,
                     missing_grad=missing_grad,
+                    defaults=defaults,
                     require_original_output_arg=require_original_output_arg,
                 )
 
@@ -1101,6 +1159,8 @@ def add_builtin(
         input_types=input_types,
         value_type=value_type,
         value_func=value_func,
+        export_func=export_func,
+        dispatch_func=dispatch_func,
         variadic=variadic,
         initializer_list_func=initializer_list_func,
         export=export,
@@ -1244,7 +1304,7 @@ class ModuleBuilder:
             if not func.value_func:
 
                 def wrap(adj):
-                    def value_type(arg_types, kwds, templates):
+                    def value_type(arg_types, arg_values):
                         if adj.return_var is None or len(adj.return_var) == 0:
                             return None
                         if len(adj.return_var) == 1:
@@ -5181,7 +5241,7 @@ def print_function(f, file, noentry=False):  # pragma: no cover
     try:
         # todo: construct a default value for each of the functions args
         # so we can generate the return type for overloaded functions
-        return_type = " -> " + type_str(f.value_func(None, None, None))
+        return_type = " -> " + type_str(f.value_func(None, None))
     except Exception:
         pass
 
@@ -5339,7 +5399,7 @@ def export_stubs(file):  # pragma: no cover
             try:
                 # todo: construct a default value for each of the functions args
                 # so we can generate the return type for overloaded functions
-                return_type = f.value_func(None, None, None)
+                return_type = f.value_func(None, None)
                 if return_type:
                     return_str = " -> " + type_str(return_type)
 
@@ -5385,20 +5445,24 @@ def export_builtins(file: io.TextIOBase):  # pragma: no cover
             if not f.is_simple():
                 continue
 
-            args = ", ".join(f"{ctype_arg_str(v)} {k}" for k, v in f.input_types.items())
-            params = ", ".join(f.input_types.keys())
-
-            return_type = ""
-
             try:
                 # todo: construct a default value for each of the functions args
                 # so we can generate the return type for overloaded functions
-                return_type = ctype_ret_str(f.value_func(None, None, None))
+                return_type = ctype_ret_str(f.value_func(None, None))
             except Exception:
                 continue
 
             if return_type.startswith("Tuple"):
                 continue
+
+            # Runtime arguments that are to be passed to the function, not its template signature.
+            if f.export_func is not None:
+                func_args = f.export_func(f.input_types)
+            else:
+                func_args = f.input_types
+
+            args = ", ".join(f"{ctype_arg_str(v)} {k}" for k, v in func_args.items())
+            params = ", ".join(func_args.keys())
 
             if args == "":
                 file.write(f"WP_API void {f.mangled_name}({return_type}* ret) {{ *ret = wp::{f.key}({params}); }}\n")