+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 ## [1.2.2] - 2024-07-04
 
 - Support for NumPy >= 2.0
+- Conform to Python's syntax for function arguments when calling built-ins inside of kernels, thus extending support for keyword arguments
 
 ## [1.2.1] - 2024-06-14